refactor: extract v3 implementation
