+++ src/test/java/io/gravitee/policy/circuitbreaker/CircuitBreakerPolicyIntegrationV3Test.java
@@ -20,24 +20,15 @@ import static com.github.tomakehurst.wiremock.client.WireMock.get;
 import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.ok;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
-import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
-import io.gravitee.apim.gateway.tests.sdk.AbstractPolicyTest;
 import io.gravitee.apim.gateway.tests.sdk.annotations.DeployApi;
-import io.gravitee.apim.gateway.tests.sdk.annotations.GatewayTest;
 import io.gravitee.common.http.HttpStatusCode;
-import io.gravitee.definition.model.Api;
-import io.gravitee.definition.model.ExecutionMode;
-import io.gravitee.policy.circuitbreaker.configuration.CircuitBreakerPolicyConfiguration;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.rxjava3.core.http.HttpClient;
 import io.vertx.rxjava3.core.http.HttpClientRequest;
 import java.util.concurrent.TimeUnit;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
 
 /**
  * Circuit breaker internal behavior is based on a registry with {@link io.gravitee.gateway.api.ExecutionContext#ATTR_RESOLVED_PATH} as key.
@@ -47,40 +38,15 @@ import org.junit.jupiter.api.extension.RegisterExtension;
  * @author Yann TAVERNIER (yann.tavernier at graviteesource.com)
  * @author GraviteeSource Team
  */
-@GatewayTest(v2ExecutionMode = ExecutionMode.V3)
-class CircuitBreakerPolicyIntegrationTest extends AbstractPolicyTest<CircuitBreakerPolicy, CircuitBreakerPolicyConfiguration> {
-
-    public static final String LOCALHOST = "localhost:";
-    public static final String REDIRECT_URL = LOCALHOST + "8089";
-
-    @RegisterExtension
-    static WireMockExtension redirectServer = WireMockExtension.newInstance().options(wireMockConfig().dynamicPort()).build();
-
-    @Override
-    public void configureApi(Api api) {
-        if (api.getId().equals("my-api-redirect")) {
-            api
-                .getFlows()
-                .forEach(flow ->
-                    flow
-                        .getPre()
-                        .stream()
-                        .filter(step -> policyName().equals(step.getPolicy()))
-                        .forEach(step ->
-                            step.setConfiguration(step.getConfiguration().replace(REDIRECT_URL, LOCALHOST + redirectServer.getPort()))
-                        )
-                );
-        }
-    }
+class CircuitBreakerPolicyIntegrationV3Test extends V3EngineTest {
 
     @Test
-    @DeployApi("/apis/circuit-breaker.json")
-    @DisplayName("Should open circuit when too many slow calls")
-    void shouldOpenCircuitWhenSlowCalls(HttpClient client) {
+    @DeployApi("/apis/v2/circuit-breaker.json")
+    void should_open_circuit_when_too_many_slow_calls(HttpClient client) {
         wiremock.stubFor(get("/endpoint/my_team").willReturn(ok("response from backend").withFixedDelay(750)));
 
         client
-            .rxRequest(HttpMethod.GET, "/test/my_team")
+            .rxRequest(HttpMethod.GET, "/v2-circuit-breaker/my_team")
             .flatMap(HttpClientRequest::rxSend)
             .test()
             .awaitDone(10, TimeUnit.SECONDS)
@@ -94,7 +60,7 @@ class CircuitBreakerPolicyIntegrationTest extends AbstractPolicyTest<CircuitBrea
         wiremock.stubFor(get("/endpoint/my_team").willReturn(ok("response from backend").withFixedDelay(750)));
 
         client
-            .rxRequest(HttpMethod.GET, "/test/my_team")
+            .rxRequest(HttpMethod.GET, "/v2-circuit-breaker/my_team")
             .flatMap(HttpClientRequest::rxSend)
             .flatMapPublisher(response -> {
                 assertThat(response.statusCode()).isEqualTo(HttpStatusCode.SERVICE_UNAVAILABLE_503);
@@ -114,13 +80,12 @@ class CircuitBreakerPolicyIntegrationTest extends AbstractPolicyTest<CircuitBrea
     }
 
     @Test
-    @DeployApi("/apis/circuit-breaker.json")
-    @DisplayName("Should open circuit when too many failures")
-    void shouldOpenCircuitWhenFailures(HttpClient client) {
+    @DeployApi("/apis/v2/circuit-breaker.json")
+    void should_open_circuit_when_too_many_failures(HttpClient client) {
         wiremock.stubFor(get("/endpoint").willReturn(aResponse().withStatus(505).withBody("response from backend")));
 
         client
-            .rxRequest(HttpMethod.GET, "/test")
+            .rxRequest(HttpMethod.GET, "/v2-circuit-breaker")
             .flatMap(HttpClientRequest::rxSend)
             .flatMapPublisher(response -> {
                 assertThat(response.statusCode()).isEqualTo(505);
@@ -138,7 +103,7 @@ class CircuitBreakerPolicyIntegrationTest extends AbstractPolicyTest<CircuitBrea
         wiremock.stubFor(get("/endpoint").willReturn(aResponse().withStatus(505).withBody("response from backend")));
 
         client
-            .rxRequest(HttpMethod.GET, "/test")
+            .rxRequest(HttpMethod.GET, "/v2-circuit-breaker")
             .flatMap(HttpClientRequest::rxSend)
             .flatMapPublisher(response -> {
                 assertThat(response.statusCode()).isEqualTo(HttpStatusCode.SERVICE_UNAVAILABLE_503);
@@ -158,13 +123,12 @@ class CircuitBreakerPolicyIntegrationTest extends AbstractPolicyTest<CircuitBrea
     }
 
     @Test
-    @DeployApi("/apis/circuit-breaker-redirect.json")
-    @DisplayName("Should redirect to URL if opened circuit due to too many failures")
-    void shouldRedirectWhenCircuitOpen(HttpClient client) {
+    @DeployApi("/apis/v2/circuit-breaker-redirect.json")
+    void should_redirect_to_url_when_circuit_opens(HttpClient client) {
         wiremock.stubFor(get("/endpoint").willReturn(aResponse().withStatus(505).withBody("response from backend")));
 
         client
-            .rxRequest(HttpMethod.GET, "/circuit-redirect")
+            .rxRequest(HttpMethod.GET, "/v2-circuit-breaker-redirect")
             .flatMap(HttpClientRequest::rxSend)
             .flatMapPublisher(response -> {
                 assertThat(response.statusCode()).isEqualTo(505);
@@ -183,7 +147,7 @@ class CircuitBreakerPolicyIntegrationTest extends AbstractPolicyTest<CircuitBrea
         redirectServer.stubFor(get("/redirection").willReturn(ok("redirection went well")));
 
         client
-            .rxRequest(HttpMethod.GET, "/circuit-redirect")
+            .rxRequest(HttpMethod.GET, "/v2-circuit-breaker-redirect")
             .flatMap(HttpClientRequest::rxSend)
             .flatMapPublisher(response -> {
                 assertThat(response.statusCode()).isEqualTo(200);
