+++ src/test/resources/apis/v2/circuit-breaker-redirect.json
@@ -1,9 +1,9 @@
 {
-    "id": "my-api",
-    "name": "my-api",
+    "id": "v2-circuit-breaker-redirect",
+    "name": "v2-circuit-breaker-redirect",
     "gravitee": "2.0.0",
     "proxy": {
-        "context_path": "/test",
+        "context_path": "/v2-circuit-breaker-redirect",
         "endpoints": [
             {
                 "name": "default",
@@ -36,7 +36,7 @@
                         "slowCallDurationThreshold": 500,
                         "windowSize": 2,
                         "waitDurationInOpenState": 50000,
-                        "redirectToURL": ""
+                        "redirectToURL": "http://localhost:8089/redirection"
                     }
                 }
             ],
